requirements.txt
Django==1.9.7
django-filter==0.13.0
django-markdown==0.8.4
djangorestframework==3.3.3
djangorestframework-gis==0.10.1
Markdown==2.6.6
psycopg2==2.6.1
